+++ apps/exams/exceptions.py
@@ -10,4 +10,3 @@ class ErrorDetailException(APIException):
         super().__init__(detail=detail)
         self.http_status = http_status
         self.status_code = http_status
-
