--- apps/exams/views/admin/questions_delete.py
@@ -1,102 +1,0 @@
-from typing import NoReturn
-
-from drf_spectacular.utils import OpenApiExample, OpenApiResponse, extend_schema
-from rest_framework import status
-from rest_framework.exceptions import NotAuthenticated, PermissionDenied
-from rest_framework.permissions import IsAuthenticated
-from rest_framework.request import Request
-from rest_framework.response import Response
-from rest_framework.views import APIView
-
-from apps.core.utils.permissions import IsStaffRole
-from apps.exams.constants import ErrorMessages
-from apps.exams.exceptions import ErrorDetailException
-from apps.exams.serializers.admin.questions_delete import (
-    AdminExamQuestionDeleteResponseSerializer,
-)
-from apps.exams.serializers.error_serializers import ErrorResponseSerializer
-from apps.exams.services.admin.questions_delete import delete_exam_question
-from apps.exams.views.mixins import ExamsExceptionMixin
-
-
-@extend_schema(
-    tags=["admin_exams"],
-    summary="어드민 문제 삭제",
-    description="쪽지시험 문제를 삭제합니다.",
-    responses={
-        200: AdminExamQuestionDeleteResponseSerializer,
-        400: OpenApiResponse(
-            response=ErrorResponseSerializer,
-            description="Bad Request",
-            examples=[
-                OpenApiExample(
-                    "유효하지 않은 문제 삭제 요청",
-                    value={"error_detail": ErrorMessages.INVALID_QUESTION_DELETE_REQUEST.value},
-                ),
-            ],
-        ),
-        401: OpenApiResponse(
-            response=ErrorResponseSerializer,
-            description="Unauthorized",
-            examples=[
-                OpenApiExample(
-                    "인증 실패",
-                    value={"error_detail": ErrorMessages.UNAUTHORIZED.value},
-                ),
-            ],
-        ),
-        403: OpenApiResponse(
-            response=ErrorResponseSerializer,
-            description="Forbidden",
-            examples=[
-                OpenApiExample(
-                    "권한 없음",
-                    value={"error_detail": ErrorMessages.NO_QUESTION_DELETE_PERMISSION.value},
-                ),
-            ],
-        ),
-        404: OpenApiResponse(
-            response=ErrorResponseSerializer,
-            description="Not Found",
-            examples=[
-                OpenApiExample(
-                    "문제 정보 없음",
-                    value={"error_detail": ErrorMessages.QUESTION_NOT_FOUND.value},
-                ),
-            ],
-        ),
-        409: OpenApiResponse(
-            response=ErrorResponseSerializer,
-            description="Conflict",
-            examples=[
-                OpenApiExample(
-                    "문제 삭제 충돌",
-                    value={"error_detail": ErrorMessages.QUESTION_DELETE_CONFLICT.value},
-                ),
-            ],
-        ),
-    },
-)
-class AdminExamQuestionDeleteAPIView(ExamsExceptionMixin, APIView):
-    """어드민 쪽지시험 문제 삭제 API."""
-
-    permission_classes = [IsAuthenticated, IsStaffRole]
-    serializer_class = AdminExamQuestionDeleteResponseSerializer
-
-    def permission_denied(self, request: Request, message: str | None = None, code: str | None = None) -> NoReturn:
-        if not request.user or not request.user.is_authenticated:
-            raise NotAuthenticated()
-        raise PermissionDenied(detail=ErrorMessages.NO_QUESTION_DELETE_PERMISSION.value)
-
-    def delete(self, request: Request, question_id: int) -> Response:
-        if question_id <= 0:
-            raise ErrorDetailException(
-                ErrorMessages.INVALID_QUESTION_DELETE_REQUEST.value,
-                status.HTTP_400_BAD_REQUEST,
-            )
-
-        result = delete_exam_question(question_id)
-
-        serializer = self.serializer_class(data=result)
-        serializer.is_valid(raise_exception=True)
-        return Response(serializer.data, status=status.HTTP_200_OK)
